refactor(analysis): rename parserdialog for all filters (#4008)
