+++ analysis/analysers/filters/EdgeFilter/src/main/kotlin/de/maibornwolff/codecharta/filter/edgefilter/Dialog.kt
@@ -9,7 +9,7 @@ import de.maibornwolff.codecharta.dialogProvider.myPromptInput
 import de.maibornwolff.codecharta.serialization.FileExtension
 import de.maibornwolff.codecharta.util.Logger
 
-class ParserDialog {
+class Dialog {
     companion object : ParserDialogInterface {
         override fun collectParserArgs(session: Session): List<String> {
             val inputFileName: String = session.myPromptDefaultFileFolderInput(
