+++ analysis/analysers/filters/StructureModifier/src/main/kotlin/de/maibornwolff/codecharta/filter/structuremodifier/Dialog.kt
@@ -10,7 +10,7 @@ import de.maibornwolff.codecharta.dialogProvider.myPromptInputNumber
 import de.maibornwolff.codecharta.dialogProvider.myPromptList
 import de.maibornwolff.codecharta.serialization.FileExtension
 
-class ParserDialog {
+class Dialog {
     companion object : ParserDialogInterface {
         override fun collectParserArgs(session: Session): List<String> {
             val inputFileName: String =
